define oclintrc rule no-restricted-imports, enforce for components, move users.ts to testhelpers
