+++ site/src/components/MultiUserSelect/MultiUserSelect.stories.tsx
@@ -2,8 +2,8 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import { spyOn } from "storybook/test";
 import { API } from "#/api/api";
 import { usersKey } from "#/api/queries/users";
-import { MockUsers } from "#/pages/UsersPage/storybookData/users";
 import { mockApiError } from "#/testHelpers/entities";
+import { MockUsers } from "#/testHelpers/users";
 import { MultiUserSelect } from "./MultiUserSelect";
 
 const meta: Meta<typeof MultiUserSelect> = {
